--- packages/hekireki/src/shared/validator/is-fields.ts
@@ -1,27 +1,0 @@
-/**
- * Extract fields with validation from a nested array of model fields.
- *
- * @param modelFields - A nested array of model field definitions.
- * @returns A flat array of fields that include a non-null `validation` property.
- */
-export function isFields(
-  modelFields: {
-    documentation: string | undefined
-    modelName: string
-    fieldName: string
-    comment: string[]
-    validation: string | null
-  }[][],
-) {
-  return modelFields.flat().filter(
-    (
-      field,
-    ): field is Required<{
-      documentation: string
-      modelName: string
-      fieldName: string
-      comment: string[]
-      validation: string | null
-    }> => field.validation !== null,
-  )
-}
